refactor(connector): streamline connector telemetry imports and clean up unused constants
